Created Type.kt to declutter AbstractSyntaxTree.kt
